telegram_bot/requirements.txt
aiogram
httpx
python-dotenv
asyncio
fastapi
uvicorn
pydantic